+++ RandomChoiceApp/View/Cells/CommonActionButtonTableViewCell.swift
@@ -81,14 +81,3 @@ class CommonActionButtonTableViewCell: UITableViewCell {
         cancelButton.layer.cornerRadius = Layer.cornerRadius
     }
 }
-
-struct Layer {
-    static let borderWidth: CGFloat = 1.0
-    static let cornerRadius: CGFloat = 28.0
-}
-
-struct Color {
-    static let white = #colorLiteral(red: 1.0, green: 1.0, blue: 1.0, alpha: 1.0)
-    static let red = #colorLiteral(red: 0.9972122312, green: 0.4152126908, blue: 0.3679206967, alpha: 1)
-    static let yellow = #colorLiteral(red: 0.9921568627, green: 0.6393200201, blue: 0.218539124, alpha: 1)
-}
